+++ src/m365/spo/commands/group/group-member-list.ts
@@ -16,13 +16,13 @@ interface Options extends GlobalOptions {
   groupName?: string;
 }
 
-class SpoGroupUserListCommand extends SpoCommand {
+class SpoGroupMemberListCommand extends SpoCommand {
   public get name(): string {
-    return commands.GROUP_USER_LIST;
+    return commands.GROUP_MEMBER_LIST;
   }
 
   public get description(): string {
-    return `List members of a SharePoint Group`;
+    return `List the members of a SharePoint Group`;
   }
 
   public defaultProperties(): string[] | undefined {
@@ -96,7 +96,7 @@ class SpoGroupUserListCommand extends SpoCommand {
 
     try {
       const response = await request.get<any>(requestOptions);
-      logger.log(response);
+      logger.log(response.value);
     }
     catch (err: any) {
       this.handleRejectedODataJsonPromise(err);
@@ -104,4 +104,4 @@ class SpoGroupUserListCommand extends SpoCommand {
   }
 }
 
-module.exports = new SpoGroupUserListCommand();
+module.exports = new SpoGroupMemberListCommand();